디자인 모달 컴포넌트로 분리 home design 추가

+++ src/components/Review/grid/ReviewModalComp.jsx
@@ -12,30 +12,29 @@ import TempReviewThumbnail from './ReviewThumbnail'
 import { useSelector } from 'react-redux'
 import { Popover, Overlay } from 'react-bootstrap'
 import { useRef, useState } from 'react'
-import { useNavigate } from 'react-router-dom'
+import { useLocation, useNavigate } from 'react-router-dom'
 import { deleteFirebaseData } from '../../../datasources/firebase'
 
-
 const ReviewModalComp = (props) => {
-  const userNow = useSelector((user)=> user.enteruser.user)
-  const navigate = useNavigate();
-  const { 
-    review, 
-    rating, 
-    tages, 
-    user, 
-    images, 
-    createdAt, 
-    boughtDate, 
-    itemName, 
-    itemColor 
+  const userNow = useSelector((user) => user.enteruser.user)
+  const navigate = useNavigate()
+  const {
+    review,
+    rating,
+    tages,
+    user,
+    images,
+    createdAt,
+    boughtDate,
+    itemName,
+    itemColor,
   } = props.review
 
   // 수정, 삭제 팝오버
-  const [ show, setShow ] = useState(false);
-  const [ target, setTarget ] = useState(null);
-  const popref = useRef(null);
-  
+  const [show, setShow] = useState(false)
+  const [target, setTarget] = useState(null)
+  const popref = useRef(null)
+
   const handleClick = (e) => {
     setShow(!show)
     setTarget(e.target)
@@ -43,14 +42,19 @@ const ReviewModalComp = (props) => {
   // 삭제버튼 클릭시
   const deletePost = async (id) => {
     try {
-        alert('정말 삭제하시겠습니까?')
+      if (window.confirm('정말 삭제하시겠습니까?') === true) {
         await deleteFirebaseData('Review', id)
-        navigate(-1)
-    } catch (e) { console.log(e) }
+        window.location.reload()
+      }
+    } catch (e) {
+      console.log(e)
+    }
   }
   // 날짜표시
-  const timeStamp = createdAt;
-  let postDate = new Date(timeStamp);
+  const timeStamp = createdAt
+  let postDate = new Date(timeStamp)
+
+  console.log(images)
 
   return (
     <div>
@@ -58,9 +62,14 @@ const ReviewModalComp = (props) => {
         button={<TempReviewThumbnail review={props.review} />}
         image={
           <SliderComp dots={false} infinite={true}>
-            { Object.values(images).map( (image,i) => (
+            {Object.values(images).map((image, i) => (
               <div key={i}>
-                <img id="image" src={image.url} key={image.name} alt="review-image" />
+                <img
+                  id="image"
+                  src={image.url}
+                  key={image.name}
+                  alt="review-image"
+                />
               </div>
             ))}
           </SliderComp>
@@ -104,7 +113,9 @@ const ReviewModalComp = (props) => {
             <div>
               <p>{user.displayName}</p>
               <p className="caption">
-                {`${postDate.getFullYear()}-${postDate.getMonth()+1}-${postDate.getDate()}`}
+                {`${postDate.getFullYear()}-${
+                  postDate.getMonth() + 1
+                }-${postDate.getDate()}`}
               </p>
             </div>
           </div>
@@ -115,12 +126,11 @@ const ReviewModalComp = (props) => {
             <ButtonComp icon>
               <FontAwesomeIcon icon={solid('share-nodes')} />
             </ButtonComp>
-            {
-              userNow && userNow.uid == user.uid ? (
-                <div>
-                  <ButtonComp icon onClick={handleClick}>
-                    <FontAwesomeIcon icon={solid("ellipsis-vertical")} />
-                  </ButtonComp>
+            {userNow && userNow.uid == user.uid ? (
+              <div>
+                <ButtonComp icon onClick={handleClick}>
+                  <FontAwesomeIcon icon={solid('ellipsis-vertical')} />
+                </ButtonComp>
                 <Overlay
                   show={show}
                   target={target}
@@ -130,18 +140,26 @@ const ReviewModalComp = (props) => {
                   rootClose
                   onHide={() => setShow(false)}
                 >
-                  <Popover id='review_popover'>
-                    <ButtonComp icon onClick={() => navigate(`/review/write/${props.review.id}`)}> 
-                      <FontAwesomeIcon icon={solid("pen-to-square")}/> 수정
-                    </ButtonComp> <br/>
-                    <ButtonComp icon onClick={()=> deletePost(props.review.id)}>
-                      <FontAwesomeIcon  icon={solid("trash-can")} /> 삭제
+                  <Popover id="review_popover">
+                    <ButtonComp
+                      color="white"
+                      onClick={() =>
+                        navigate(`/review/write/${props.review.id}`)
+                      }
+                    >
+                      <FontAwesomeIcon icon={solid('pen-to-square')} /> 수정
+                    </ButtonComp>{' '}
+                    <br />
+                    <ButtonComp
+                      color="white"
+                      onClick={() => deletePost(props.review.id)}
+                    >
+                      <FontAwesomeIcon icon={solid('trash-can')} /> 삭제
                     </ButtonComp>
                   </Popover>
                 </Overlay>
               </div>
-              ) : null
-            }
+            ) : null}
           </div>
         </div>
       </ModalComp>
